HS-210 shared libray for REST server and platform core
