+++ platform/alarms/rest/pom.xml
@@ -56,9 +56,9 @@
       <artifactId>swagger-annotations</artifactId>
     </dependency>
     <dependency>
-      <groupId>org.opennms.horizon.db</groupId>
-      <artifactId>common-model</artifactId>
-      <version>0.1.0-SNAPSHOT</version>
+      <groupId>org.opennms.horizon.shared</groupId>
+      <artifactId>dto-model</artifactId>
+      <version>${horizon.shared.lib.version}</version>
     </dependency>
   </dependencies>
