+++ platform/alarms/rest/src/main/java/org/opennms/netmgt/alarmd/rest/internal/AlarmRestServiceImpl.java
@@ -58,9 +58,6 @@ import javax.ws.rs.core.UriInfo;
 
 import org.apache.commons.lang3.EnumUtils;
 import org.apache.commons.lang3.StringUtils;
-import org.opennms.horizon.db.common.model.AlarmAckDTO;
-import org.opennms.horizon.db.common.model.AlarmCollectionDTO;
-import org.opennms.horizon.db.common.model.AlarmDTO;
 import org.opennms.horizon.core.lib.SystemProperties;
 import org.opennms.horizon.db.dao.api.AcknowledgmentDao;
 import org.opennms.horizon.db.dao.api.AlarmDao;
@@ -71,6 +68,9 @@ import org.opennms.horizon.db.model.OnmsAcknowledgment;
 import org.opennms.horizon.db.model.OnmsAlarm;
 import org.opennms.horizon.db.model.TroubleTicketState;
 import org.opennms.horizon.db.model.mapper.AlarmMapper;
+import org.opennms.horizon.shared.dto.AlarmAckDTO;
+import org.opennms.horizon.shared.dto.AlarmCollectionDTO;
+import org.opennms.horizon.shared.dto.AlarmDTO;
 import org.opennms.netmgt.alarmd.rest.AlarmRestService;
 import org.opennms.web.rest.support.MultivaluedMapImpl;
 import org.opennms.web.rest.support.SecurityHelper;
