+++ platform/events/api/src/main/java/org/opennms/horizon/events/xml/DateTimeAdapter.java
@@ -47,5 +47,4 @@ public class DateTimeAdapter extends XmlAdapter<String, Date> {
     public Date unmarshal(final String string) throws Exception {
         return (string == null || string.isEmpty()) ? null : EventConstants.parseToDate(string);
     }
-
 }
