+++ platform/pom.xml
@@ -55,6 +55,7 @@
         <hibernate.version>5.4.32.Final</hibernate.version>
         <httpcore.version>4.4.4</httpcore.version>
         <httpclient.version>4.5.13</httpclient.version>
+        <horizon.shared.lib.version>0.1.0-SNAPSHOT</horizon.shared.lib.version>
         <jackson.version>2.13.1</jackson.version>
         <jaeger.version>0.34.0</jaeger.version>
         <jakarta.annotation.version>1.3.5</jakarta.annotation.version>
@@ -241,6 +242,16 @@
             <id>opennms-snapshots</id>
             <name>OpenNMS Snapshot Maven Repository</name>
             <url>https://maven.opennms.org/content/groups/opennms.org-snapshot</url>
+        </repository>
+        <repository>
+            <id>opennms-maven-3rdparty</id>
+            <url>https://packages.opennms.com/public/3rdparty/maven/</url>
+            <releases>
+                <enabled>true</enabled>
+            </releases>
+            <snapshots>
+                <enabled>true</enabled>
+            </snapshots>
         </repository>
     </repositories>
     <profiles>
