+++ shared-lib/dto-model/src/main/java/org/opennms/horizon/shared/dto/AlarmAckDTO.java
@@ -1,4 +1,4 @@
-package org.opennms.horizon.db.common.model;
+package org.opennms.horizon.shared.dto;
 
 public class AlarmAckDTO {
     private String user;
