+++ shared-lib/dto-model/src/main/java/org/opennms/horizon/shared/dto/EventDTO.java
@@ -26,7 +26,7 @@
  *     http://www.opennms.com/
  *******************************************************************************/
 
-package org.opennms.horizon.db.common.model;
+package org.opennms.horizon.shared.dto;
 
 import java.util.ArrayList;
 import java.util.Date;
